Add method to insert minimal dependency in FunctionalDependencies
This commit includes the addition of the method 'insertMinimalDependency'. Its purpose is to efficiently insert minimal dependencies into the existing data structure. This is done by checking for subsets and supersets of the new dependency and modifying existing entries where necessary. This will improve the performance of handling minimal dependencies in our application.