Add maven-shade-plugin and update module names
In the pom.xml files, the maven-shade-plugin has been added to configure the creation of an uber-jar. The primary class is now specified within this plugin. In addition, the module name 'DFDModules' has been updated to 'DDFDModules' and the same changes have been applied to the class path to align with the updated module names. A new module 'ddfd' has also been added to the root pom.xml file. These changes ensure better package integrity and clarity in module distinction.