Add check to prevent endless seed generation in GraphTraverser
Introduced some lines in GraphTraverser.java to keep track of previously generated seeds. This will help to assure that the code does not enter into an infinite loop when generating seeds. It also asserts that the new seeds generated in each iteration are not a repetition of the previous ones, thus avoiding unnecessary computations and potential infinite loops.