Add statistics tracking to partitions
This commit adds an option to track statistics in the Partition and ComposedPartition classes. The recreation count for each column collection is now incremented each time a new partition is instantiated with the track statistics option. The results are then printed out at the end of the DFD miner's run method. This information will help to measure the performance of the partitioning algorithm and identify potential areas for optimization.