Add feature to display mined Functional Dependencies
A new functionality to display mined Functional Dependencies (FDs) has been added. This was achieved by introducing a new static variable, SHOW_FDS, which is initially set to false. A new case within the main function's command-line argument parsing section has been added, in which the developer can set --showfds to true. If SHOW_FDS is set to true, all the mined FDs will be printed on the console. The documentation for the command-line interface has also been updated to reflect this change.